+++ server_disable_rescue.go
@@ -33,10 +33,11 @@ func runServerDisableRescue(cli *CLI, cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return err
 	}
-	errCh, _ := waitAction(cli.Context, cli.Client(), action)
-	if err := <-errCh; err != nil {
+
+	if err := cli.ActionProgress(cli.Context, action); err != nil {
 		return err
 	}
+
 	fmt.Printf("Rescue disabled for server %s\n", server.Name)
 	return nil
 }