+++ server_rebuild.go
@@ -53,10 +53,11 @@ func runServerRebuild(cli *CLI, cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return err
 	}
-	errCh, _ := waitAction(cli.Context, cli.Client(), action)
-	if err := <-errCh; err != nil {
+
+	if err := cli.ActionProgress(cli.Context, action); err != nil {
 		return err
 	}
+
 	fmt.Printf("Server %s rebuilt with image %s\n", server.Name, imageIDOrName)
 	return nil
 }